Add breakable dungeon props (barrels, crates) with server sync
Clicking a barrel or crate in a dungeon makes the player walk up,
swing the sword once, and replace the whole-model GLB with its
broken-debris variant (barrel→barrel_broken, crate→crate_pieces)
at the contact frame. Chests are not breakable.

Server tracks broken props in DungeonRuntime.broken_props, validates
floor/proximity/kind, broadcasts DungeonPropBroken to nearby players
via floor-gated AOI, and sends a DungeonPropsState snapshot on floor
entry so late/re-entering players see current state. In-memory only
(resets on server restart, matching chest-cooldown behaviour).

The broken prop's cell is made walkable: a new dungeon_apply_broken_props
wasm binding rebuilds the passability cache for the floor by recomputing
from the carved mask, so walls the prop sat against stay blocked.

New protocol messages: ClientMessage::BreakDungeonProp,
ServerMessage::DungeonPropBroken / DungeonPropsState.

Client: break_prop raycast pass + intent, walk-up-then-break via a
pending-break that the dungeon layer fires on proximity, one-shot sword
swing reusing the existing 'attack' FSM state. Shared buildBrokenClone
helper extracted; timer cleanup added on unmount.

Fix: dungeon gen used gen_range(0..slice.len()) in roll_spawns and
roll_props; usize is u32 on wasm32 but u64 on native, so the RNG drew
differently per platform — client (wasm) and server (native) disagreed
on prop cells, causing break validation to reject "too far". Fixed both
to a fixed-width u32 draw; golden-layout-hash test re-blessed.

+++ shared/src/dungeon/gen.rs
@@ -440,7 +440,11 @@ fn roll_spawns(rng: &mut ChaCha8Rng, layout: &FloorLayout) -> Vec<SpawnSpec> {
     };
 
     for _ in 0..count.min(candidates.len()) {
-        let idx = rng.gen_range(0..candidates.len());
+        // `as u32`: `usize` is u32 on wasm32 but u64 on native, and
+        // `gen_range` over each width consumes a different amount of the RNG
+        // stream — desyncing client (wasm) from server (native) here, and in
+        // every draw after. Fixed-width keeps the layout identical on both.
+        let idx = rng.gen_range(0..candidates.len() as u32) as usize;
         let (x, z) = candidates.swap_remove(idx);
         let mut roll = rng.gen_range(0..total_weight);
         let mut chosen = &table[0];
@@ -609,7 +613,11 @@ fn roll_props(rng: &mut ChaCha8Rng, layout: &FloorLayout) -> Vec<PropSpec> {
             if pool.is_empty() {
                 break;
             }
-            let (x, z) = pool.swap_remove(rng.gen_range(0..pool.len()));
+            // `as u32`: see roll_spawns — `gen_range` over `usize` (u32 on
+            // wasm, u64 on native) draws differently per platform, desyncing the
+            // RNG. This is the draw whose divergence shows up as props sitting in
+            // different cells on the client vs the server.
+            let (x, z) = pool.swap_remove(rng.gen_range(0..pool.len() as u32) as usize);
             let idx = (x + z * GRID) as usize;
             taken[idx] = true;
